Upgrade to springboot3

+++ agileboot-domain/src/main/java/com/agileboot/domain/system/monitor/dto/CpuInfo.java
@@ -1,7 +1,7 @@
 package com.agileboot.domain.system.monitor.dto;
 
-import cn.hutool.core.util.NumberUtil;
 import lombok.Data;
+import org.dromara.hutool.core.math.NumberUtil;
 
 /**
  * CPU相关信息
@@ -46,19 +46,19 @@ public class CpuInfo {
     }
 
     public double getSys() {
-        return NumberUtil.div(sys * 100, total, 2);
+        return NumberUtil.div(sys * 100, total, 2).doubleValue();
     }
 
     public double getUsed() {
-        return NumberUtil.div(used * 100, total, 2);
+        return NumberUtil.div(used * 100, total, 2).doubleValue();
     }
 
     public double getWait() {
-        return NumberUtil.div(wait * 100, total, 2);
+        return NumberUtil.div(wait * 100, total, 2).doubleValue();
     }
 
     public double getFree() {
-        return NumberUtil.div(free * 100, total, 2);
+        return NumberUtil.div(free * 100, total, 2).doubleValue();
     }
 }
